+++ packages/components/components/elvis-dropdown/src/react/declarations.d.ts
@@ -1,1 +1,3 @@
 declare module '*';
+declare module '@elvia/elvis-colors';
+declare module '@elvia/elvis-typography';