+++ packages/web/src/app/doc-pages/components/modal-doc/modal-info-code.ts
@@ -39,7 +39,7 @@ const modalInfoCode = {
 </elvia-modal>
 `,
   codeNativeScript: ``,
-// all scripts for the examples are in modal-data.ts
+  // all scripts for the examples are in modal-data.ts
 };
 
 export { modalInfoCode };