+++ packages/components/components/elvis-accordion/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Elvia Accordion Changelog
 
+## 2.0.6 (02.12.21)
+
+### Patch
+
+- Updated and added more tests
+
+## 2.0.5 (30.11.21)
+
+### Patch
+
+- Now using elvis-colors package for colors.
+
+## 2.0.4 (30.11.21)
+
+- Web component wrapper fix
+
+### Bug fix
+
+- Improved type safety.
+
 ## 2.0.3 (17.11.21)
 
 ### Bug fix