+++ packages/components/components/elvis-pagination/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Elvia Pagination Changelog
 
+## 1.0.6 (30.11.21)
+
+### Bug fix
+
+- Now using elvis-colors package for colors.
+- Now using elvis-typography package for typography.
+
+## 1.0.5 (30.11.21)
+
+### Bug fix
+
+- Web component wrapper fix
+
+## 1.0.4 (29.11.21)
+
+### Patch
+
+Pagination build with new wrapper version
+
 ## 1.0.3 (17.11.21)
 
 ### Bug fix