+++ packages/components/components/elvis-progress-linear/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Elvia Progress-linear Changelog
 
+## 1.1.1 (30.11.21)
+
+- Web component wrapper fix
+
 ## 1.1.0 (03.11.21)
 
 ### New features